feat(pando cli): cleanup
