+++ packages/pando/src/shared/types.mts
@@ -1,7 +1,5 @@
 export type CLIOperation = 'setup' | 'update' | undefined
 
-// export type PMOptions = 'bun' | 'pnpm' | 'yarn' | 'npm' | undefined
-
 export enum PMOptions {
   BUN = 'bun',
   PNPM = 'pnpm',
