+++ tests/pando/pandoUpdate.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'bun:test'
-import { setup, DOWN, ENTER } from './helpers'
+import { setup } from './helpers'
 
 describe('pando update', () => {
   test('should execute the cli with the update arg', async () => {
@@ -7,14 +7,5 @@ describe('pando update', () => {
     stdin.end()
     const res = await new Response(stdout).text()
     expect(res).toInclude('Welcome to Pando update\n')
-  })
-  test('should initiate pandoUpdate when selected', async () => {
-    const proc = setup()
-    proc.stdin.write(DOWN)
-    proc.stdin.write(ENTER)
-    proc.stdin.end()
-
-    const text = await new Response(proc.stdout).text()
-    expect(text).toInclude('Welcome to Pando update')
   })
 })
